[Fix] PencilShop 관련 Repository PencilShopRepository로 통합
